backend/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
google-auth==2.37.0
gspread==6.1.4
beautifulsoup4==4.12.3
python-dotenv==1.0.1
